+++ perf/WeihanLi.Common.Benchmark/Program.cs
@@ -16,9 +16,9 @@ namespace WeihanLi.Common.Benchmark
             serviceCollection.AddSingleton<IConfiguration>(configurationBuilder.Build());
             DependencyResolver.SetDependencyResolver(serviceCollection);
 
-            // BenchmarkRunner.Run<MapperTest>();
+            BenchmarkRunner.Run<MapperTest>();
             // BenchmarkRunner.Run<CreateInstanceTest>();
-            BenchmarkRunner.Run<DITest>();
+            // BenchmarkRunner.Run<DITest>();
 
             Console.ReadLine();
         }