+++ src/WeihanLi.Common/Extensions/QueryableExtension.cs
@@ -10,14 +10,61 @@ namespace WeihanLi.Extensions
 {
     public static class QueryableExtension
     {
+        public static IQueryable<T> WhereIf<T>(this IQueryable<T> source, Expression<Func<T, bool>> predicate, bool condition)
+        {
+            return condition ? source?.Where(predicate) : source;
+        }
+
+        public static IQueryable<T> WhereIf<T>(this IQueryable<T> source, Expression<Func<T, bool>> predicate, Func<bool> conditionFunc)
+        {
+            return conditionFunc() ? source?.Where(predicate) : source;
+        }
+
         /// <summary>
-        /// Converts the specified source to <see cref="IPagedListModel{T}"/> by the specified <paramref name="pageNumber"/> and <paramref name="pageSize"/>.
+        /// Converts the specified source to <see cref="IPagedListResult{T}"/> by the specified <paramref name="pageNumber"/> and <paramref name="pageSize"/>.
         /// </summary>
         /// <typeparam name="T">The type of the source.</typeparam>
         /// <param name="source">The source to paging.</param>
         /// <param name="pageNumber">The number of the page, start from 1.</param>
         /// <param name="pageSize">The size of the page.</param>
-        /// <returns>An instance of  implements <see cref="IPagedListModel{T}"/> interface.</returns>
+        /// <returns>An instance of  implements <see cref="IPagedListResult{T}"/> interface.</returns>
+        public static IListResultWithTotal<T> ToListResultWithTotal<T>([NotNull] this IQueryable<T> source, int pageNumber, int pageSize)
+        {
+            if (pageNumber <= 0)
+            {
+                pageNumber = 1;
+            }
+            if (pageSize <= 0)
+            {
+                pageSize = 10;
+            }
+            var count = source.Count();
+            if (count == 0)
+            {
+                return ListResultWithTotal<T>.Empty;
+            }
+
+            if (pageNumber > 1)
+            {
+                source = source.Skip((pageNumber - 1) * pageSize);
+            }
+            var items = source.Take(pageSize).ToArray();
+
+            return new ListResultWithTotal<T>()
+            {
+                TotalCount = count,
+                Data = items
+            };
+        }
+
+        /// <summary>
+        /// Converts the specified source to <see cref="IPagedListResult{T}"/> by the specified <paramref name="pageNumber"/> and <paramref name="pageSize"/>.
+        /// </summary>
+        /// <typeparam name="T">The type of the source.</typeparam>
+        /// <param name="source">The source to paging.</param>
+        /// <param name="pageNumber">The number of the page, start from 1.</param>
+        /// <param name="pageSize">The size of the page.</param>
+        /// <returns>An instance of  implements <see cref="IPagedListResult{T}"/> interface.</returns>
         public static IPagedListResult<T> ToPagedList<T>([NotNull] this IQueryable<T> source, int pageNumber, int pageSize)
         {
             if (pageNumber <= 0)