+++ src/WeihanLi.Common/IScope.cs
@@ -5,4 +5,13 @@ namespace WeihanLi.Common
     public interface IScope : IDisposable
     {
     }
+
+    public class NullScope : IScope
+    {
+        public void Dispose()
+        {
+        }
+
+        public static NullScope Instance { get; } = new NullScope();
+    }
 }