+++ test/WeihanLi.Common.Test/ExtensionsTest/StringExtensionTest.cs
@@ -122,6 +122,10 @@ namespace WeihanLi.Common.Test.ExtensionsTest
             array = str.SplitArray<int>(new[] { ';' });
             Assert.Equal(count, array.Length);
             Assert.True(array.SequenceEqual(Enumerable.Range(1, count)));
+
+            var array1 = str.SplitArray<int?>(new[] { ';' });
+            Assert.Equal(count, array1.Length);
+            Assert.True(array1.Select(x => x.GetValueOrDefault()).SequenceEqual(Enumerable.Range(1, count)));
         }
     }
 }